requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
exceptiongroup==1.3.0
fonttools==4.58.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
holidays==0.73
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.4.4
narwhals==1.41.0
numba==0.61.2
numpy==2.2.6
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.1.2
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
prophet==1.1.7
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
realtime==2.4.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
stanio==0.5.1
storage3==0.11.3
streamlit==1.45.1
streamlit-aggrid==1.1.5
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
urllib3==2.4.0
websockets==14.2
yarl==1.20.0
